extension: Move storage functions to util.js
