+++ web/packages/extension/manifest.json
@@ -14,7 +14,7 @@
     "content_scripts": [
         {
             "matches": ["<all_urls>"],
-            "js": ["js/lv0.js"],
+            "js": ["dist/lv0.js"],
             "all_frames": true,
             "run_at": "document_start"
         }
