+++ web/packages/extension/webpack.config.js
@@ -17,6 +17,7 @@ module.exports = (env, argv) => {
             ruffle: path.resolve(__dirname, "js/index.js"),
             popup: path.resolve(__dirname, "js/popup.js"),
             settings: path.resolve(__dirname, "js/settings.js"),
+            lv0: path.resolve(__dirname, "js/lv0.js"),
         },
         output: {
             path: path.resolve(__dirname, "build/dist"),
